+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# v6.4.0 (Unpublished)
+# v6.4.0
 ### New features
  - add user types to the user. All existing users will be set of type `user`
  - update the Group management description